Rename TaskTracker to TaskQueue
Also Rename `singlethreaded` package to `tasktracker`

Together with https://github.com/bluerogue251/DBSubsetter/pull/68, this brings
us closer to sharing the same "TaskQueue" interface between the Single-Threaded
and the Akka-Streams versions of subsetting. The end-goal is that the Single-
Threaded and Akka-Streams versions will eventually be able to share a single
implementation for instrumenting the "TaskQueue" with a Prometheus Gauge to count
currently outstanding tasks.
