+++ src/main/scala/trw/dbsubsetter/taskqueue/TaskQueue.scala
@@ -1,8 +1,8 @@
-package trw.dbsubsetter.singlethreaded
+package trw.dbsubsetter.taskqueue
 
 import trw.dbsubsetter.workflow.OriginDbRequest
 
-trait TaskTracker {
+trait TaskQueue {
   def nonEmpty: Boolean
   // `IndexedSeq` guarantees calls to `tasks.length()` will be O(1)
   def enqueueTasks(tasks: IndexedSeq[OriginDbRequest]): Unit
